update security configuration to user endpoint

+++ src/main/kotlin/dev/janssenbatista/shopping/list/api/services/UserService.kt
@@ -1,9 +1,11 @@
 package dev.janssenbatista.shopping.list.api.services
 
+import dev.janssenbatista.shopping.list.api.exceptions.UnauthorizedException
 import dev.janssenbatista.shopping.list.api.exceptions.UserAlreadyExistsException
 import dev.janssenbatista.shopping.list.api.exceptions.UserNotFoundException
 import dev.janssenbatista.shopping.list.api.models.User
 import dev.janssenbatista.shopping.list.api.repositories.UserRepository
+import dev.janssenbatista.shopping.list.api.secutiry.AuthenticationFacade
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.security.core.userdetails.UserDetailsService
 import org.springframework.security.core.userdetails.UsernameNotFoundException
@@ -13,12 +15,15 @@ import java.time.LocalDateTime
 import java.util.*
 
 @Service
-class UserService(private val userRepository: UserRepository) : UserDetailsService {
+class UserService(private val userRepository: UserRepository, private val authenticationFacade: AuthenticationFacade) : UserDetailsService {
 
     fun findUserById(id: UUID): User {
-        return userRepository.findById(id).orElseThrow {
+        val user = userRepository.findById(id).orElseThrow {
             throw UserNotFoundException("User with id $id not found")
         }
+        if (user.email != authenticationFacade.getAuthentication().name)
+            throw UnauthorizedException()
+        return user
     }
 
     fun save(user: User): User {
@@ -32,22 +37,25 @@ class UserService(private val userRepository: UserRepository) : UserDetailsServi
     }
 
     fun update(id: UUID, user: User): User {
-        val userAlreadyExists = userRepository.findById(id)
-        if (!userAlreadyExists.isPresent) {
-            throw UserNotFoundException("User with id $id not found")
+        val userExists = userRepository.findById(id).orElseThrow {
+            UserNotFoundException("User with id $id not found")
         }
+        if (user.email != authenticationFacade.getAuthentication().name)
+            throw UnauthorizedException()
         val hashPassword = BCryptPasswordEncoder(12).encode(user.userPassword)
-        userAlreadyExists.get().apply {
+        userExists.apply {
             email = user.email
             userPassword = hashPassword
             updatedAt = LocalDateTime.now()
         }
-        return userRepository.save(user)
+        return userRepository.save(userExists)
     }
 
     fun deleteById(id: UUID) {
-        userRepository.findById(id)
+        val user = userRepository.findById(id)
                 .orElseThrow { throw UserNotFoundException("User with id $id not found") }
+        if (user.email != authenticationFacade.getAuthentication().name)
+            throw UnauthorizedException()
         userRepository.deleteById(id)
     }
 
@@ -55,6 +63,7 @@ class UserService(private val userRepository: UserRepository) : UserDetailsServi
         return userRepository.findByEmail(email!!)
                 .orElseThrow { throw UsernameNotFoundException("User not found") }
     }
+
 }
